feat(udr): logging level can be setup at run-time and not build-time

+++ src/api_server/impl/AMF3GPPAccessRegistrationDocumentApiImpl.cpp
@@ -74,7 +74,7 @@ void AMF3GPPAccessRegistrationDocumentApiImpl::create_amf_context3gpp(
       ueId, amf3GppAccessRegistration, response_data, http_code);
 
   code = static_cast<Pistache::Http::Code>(http_code);
-  Logger::udr_server().debug("HTTP Response code %d.\n", code);
+  Logger::udr_server().debug("HTTP Response code %d.\n", (int) code);
   // content type
   response.headers().add<Pistache::Http::Header::ContentType>(
       Pistache::Http::Mime::MediaType("application/json"));
@@ -92,7 +92,7 @@ void AMF3GPPAccessRegistrationDocumentApiImpl::query_amf_context3gpp(
   m_udr_app->handle_query_amf_context_3gpp(ueId, response_data, http_code);
 
   code = static_cast<Pistache::Http::Code>(http_code);
-  Logger::udr_server().debug("HTTP Response code %d.\n", code);
+  Logger::udr_server().debug("HTTP Response code %d.\n", (int) code);
   // content type
   response.headers().add<Pistache::Http::Header::ContentType>(
       Pistache::Http::Mime::MediaType("application/json"));
