+++ src/api_server/impl/SDMSubscriptionsCollectionApiImpl.cpp
@@ -57,7 +57,7 @@ void SDMSubscriptionsCollectionApiImpl::create_sdm_subscriptions(
       ueId, sdmSubscription, response_data, http_code);
 
   code = static_cast<Pistache::Http::Code>(http_code);
-  Logger::udr_server().debug("HTTP Response code %d.\n", code);
+  Logger::udr_server().debug("HTTP Response code %d.\n", (int) code);
   // content type
   response.headers().add<Pistache::Http::Header::ContentType>(
       Pistache::Http::Mime::MediaType("application/json"));
@@ -74,7 +74,7 @@ void SDMSubscriptionsCollectionApiImpl::querysdmsubscriptions(
   m_udr_app->handle_query_sdm_subscriptions(ueId, response_data, http_code);
 
   code = static_cast<Pistache::Http::Code>(http_code);
-  Logger::udr_server().debug("HTTP Response code %d.\n", code);
+  Logger::udr_server().debug("HTTP Response code %d.\n", (int) code);
   // content type
   response.headers().add<Pistache::Http::Header::ContentType>(
       Pistache::Http::Mime::MediaType("application/json"));
